requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
configparser==5.2.0
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
flask-heroku==0.1.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
pipenv==2022.1.8
platformdirs==2.4.1
protobuf==3.19.4
psycopg2==2.9.3
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
